Add files via upload
A simple convolutional neural network that uses pytorch. The input X is a tensor, with a kernel (K) that goes over the input and provides an output (y) which represents how well the model predictions match the expected output. 

As for training the CNN, we run forward pass first where Y_hat is computes the predicted output. The loss is then calculated, leading to backward pass which computes the gradients with respect to the kernel weight. Finally, updating the kernel weight using the gradients from the learning rate.